Simplify loss tests
